python-shopping-service/requirements.txt
flask==2.0.1
flask-cors==3.0.10
selenium==4.18.1
beautifulsoup4==4.12.3
webdriver-manager==4.0.1
python-dotenv==1.0.0 